refactor: move showcase into design-system
